pyproject.toml
[project]
name = "apple-leaf-disease-classifier"
version = "0.1.0"
description = "Apple leaf desease classification from images"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dvc[s3]>=3.65.0",
    "hydra-core>=1.3.2",
    "mlflow>=3.8.1",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pre-commit>=4.5.1",
    "pytorch-lightning>=2.6.0",
    "ruff>=0.14.10",
    "scikit-learn>=1.7.2",
    "torch>=2.9.1",
    "torchmetrics>=1.8.2",
    "torchvision>=0.24.1",
]

[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = [
  ".venv",
  ".dvc",
  "data",
  "outputs",
]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N"]
ignore = ["N999"]
fixable = ["ALL"]

[tool.ruff.lint.isort]
known-first-party = ["apple_leaf_disease"]
combine-as-imports = true

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "lf"
